requirements.txt
asgiref==3.5.2
certifi @ file:///C:/b/abs_ac29jvt43w/croot/certifi_1665076682579/work/certifi
Django==4.1.4
djangorestframework==3.14.0
pytz==2022.6
sqlparse==0.4.3
tzdata==2022.7
wincertstore==0.2

pathlib~=1.0.1
ocrmypdf~=14.0.2
pandas~=1.5.2
fitz~=0.0.1.dev2
numpy~=1.23.5
gensim~=4.3.0
apscheduler~=3.9.1.post1